Builds the front-end and hosts it inside docker compose (#20)
