+++ .gitignore
@@ -267,3 +267,5 @@ poetry.toml
 pyrightconfig.json
 
 # End of https://www.toptal.com/developers/gitignore/api/django,python
+
+yivi_portal/static/
