+++ Dockerfile.django
@@ -33,7 +33,7 @@ RUN mkdir -p $DJANGO_MEDIA_ROOT
 ENV DJANGO_STATIC_URL=/static/
 ENV DJANGO_MEDIA_URL=/media/
 
-RUN python manage.py collectstatic --noinput
+# RUN python manage.py collectstatic --noinput
 
 RUN chown -R nobody:nogroup $DJANGO_MEDIA_ROOT
