+++ portal_frontend/src/pages/login/index.tsx
@@ -1,27 +1,27 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 "use client";
 
-import useStore from "@/store";
+import useStore from "@/src/store";
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
+import getConfig from "next/config";
 
 export default function Login() {
+  const { publicRuntimeConfig } = getConfig();
   const accessToken = useStore((state) => state.accessToken);
   const setAccessToken = useStore((state) => state.setAccessToken)
   const router = useRouter();
 
-  let web: any = null;
-
   useEffect(() => {
     import("@privacybydesign/yivi-frontend").then((yivi: any) => {
-      web = yivi.newWeb({
+      const web = yivi.newWeb({
         debugging: false,            // Enable to get helpful output in the browser console
         element:   '#yivi-web-form', // Which DOM element to render to
       
         // Back-end options
         session: {
           // Point this to your controller:
-          url: 'http://localhost:8000/v1',
+          url: publicRuntimeConfig.API_ENDPOINT + '/v1',
       
           start: {
             url: (o: any) => `${o.url}/session/`,
@@ -53,7 +53,7 @@ export default function Login() {
           alert(err);
         });
     });
-  }, []);
+  }, [router, setAccessToken]);
 
   return (
     <div className="flex justify-center items-center">
