+++ portal_frontend/src/pages/organizations/index.tsx
@@ -1,12 +1,14 @@
 "use client";
 
 import { useState, useEffect } from 'react';
-import axios from 'axios';
 import Link from 'next/link';
-import { Table, TableBody, TableCaption, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
+import { Table, TableBody, TableCaption, TableCell, TableHead, TableHeader, TableRow } from "@/src/components/ui/table";
+import { Badge } from "@/src/components/ui/badge";
+import { Button } from "@/src/components/ui/button";
+import { Checkbox } from "@/src/components/ui/checkbox";
+import Image from "next/image";
+import { axiosInstance } from '@/src/services/axiosInstance';
+import getConfig from 'next/config';
 
 // Define Organization type
 interface Organization {
@@ -29,6 +31,7 @@ interface PaginationResponse {
 }
 
 export default function OrganizationsPage() {
+  const { publicRuntimeConfig } = getConfig();
   const [organizations, setOrganizations] = useState<Organization[]>([]);
   const [loading, setLoading] = useState(true);
   const [trustModels, setTrustModels] = useState<string[]>([]);
@@ -50,7 +53,7 @@ export default function OrganizationsPage() {
     const offset = (page - 1) * pageSize;
     
     // Base URL with pagination
-    let url = `http://0.0.0.0:8000/v1/organizations/?limit=${pageSize}&offset=${offset}`;
+    let url = `/v1/organizations/?limit=${pageSize}&offset=${offset}`;
     
     // Only add filters if applyFilters is true
     if (applyFilters) {
@@ -67,7 +70,7 @@ export default function OrganizationsPage() {
     
     console.log("Fetching organizations:", url);
     
-    axios.get(url)
+    axiosInstance.get(url)
       .then(response => {
         const data = response.data as PaginationResponse;
         let orgs = data.results;
@@ -313,13 +316,15 @@ export default function OrganizationsPage() {
                     <TableCell>
                       {org.logo ? (
                         <div className="relative h-8 w-8 rounded-full overflow-hidden border border-gray-200">
-                          <img 
-                            src={`http://0.0.0.0:8000${org.logo}`} 
+                          <Image 
+                            src={`${publicRuntimeConfig.API_ENDPOINT}${org.logo}`}
+                            width={32}
+                            height={32}
                             alt={`${org.name_en} logo`} 
                             className="object-cover w-full h-full"
-                            onError={(e) => {
-                              e.currentTarget.src = "/placeholder-logo.png"; 
-                            }}
+                            // onError={(e) => {
+                            //   // e.currentTarget.src = "/placeholder-logo.png"; 
+                            // }}
                           />
                         </div>
                       ) : (
