+++ portal_frontend/src/services/axiosInstance.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+import getConfig from 'next/config';
+const { publicRuntimeConfig } = getConfig();
+
+export const axiosInstance = axios.create({
+  baseURL: publicRuntimeConfig.API_ENDPOINT,
+});
