+++ portal_frontend/src/styles/globals.css
