+++ yivi_portal/urls.py
@@ -9,4 +9,5 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("portal_backend.urls")),
     path("", include("yivi_auth.urls")),
+    # Serve media files separately
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
